fix it up a bit more
